# server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "区块链房地产代币化系统 API",
    "version": "1.0.0",
    "description": "测试配置"
  },
  "paths": {
    "/api/v1/core/property-manager/properties": {
      "get": {
        "summary": "获取所有房产ID",
        "description": "获取系统中所有已注册的房产ID列表",
        "tags": [
          "PropertyManager"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "房产ID列表获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/core/property-manager/properties/{propertyId}": {
      "get": {
        "summary": "获取房产详情",
        "description": "获取指定房产ID的详细信息",
        "tags": [
          "PropertyManager"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "propertyId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "房产ID"
          }
        ],
        "responses": {
          "200": {
            "description": "房产详情获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "404": {
            "description": "房产不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/core/property-manager/properties/status/{status}": {
      "get": {
        "summary": "根据状态获取房产",
        "description": "获取指定状态的所有房产ID",
        "tags": [
          "PropertyManager"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "status",
            "required": true,
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1,
                2,
                3,
                4
              ]
            },
            "description": "房产状态"
          }
        ],
        "responses": {
          "200": {
            "description": "房产列表获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/core/property-manager/owner/{ownerAddress}/properties": {
      "get": {
        "summary": "获取所有者的房产",
        "description": "获取指定所有者拥有的所有房产ID",
        "tags": [
          "PropertyManager"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "ownerAddress",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "所有者地址"
          }
        ],
        "responses": {
          "200": {
            "description": "房产列表获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/core/property-manager/update-property-status": {
      "put": {
        "summary": "更新房产状态",
        "description": "更新指定房产的状态，需要MANAGER权限",
        "tags": [
          "PropertyManager"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "propertyId",
                  "status"
                ],
                "properties": {
                  "propertyId": {
                    "type": "string",
                    "description": "房产ID"
                  },
                  "status": {
                    "type": "integer",
                    "enum": [
                      0,
                      1,
                      2,
                      3,
                      4
                    ],
                    "description": "新的房产状态"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "房产状态更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "参数错误或合约错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/core/property-manager/transfer-ownership": {
      "post": {
        "summary": "转移房产所有权",
        "description": "将房产所有权转移给新的所有者，需要当前所有者权限",
        "tags": [
          "PropertyManager"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "propertyId",
                  "newOwner"
                ],
                "properties": {
                  "propertyId": {
                    "type": "string",
                    "description": "房产ID"
                  },
                  "newOwner": {
                    "type": "string",
                    "description": "新所有者的地址"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "所有权转移成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "参数错误或合约错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}